Implemented zs's TODO list

+++ commands/Main/meta.js
@@ -2,6 +2,7 @@ module.exports = {
     name: 'meta',
     description: `Create a new category with a meta channel.`,
     aliases: ['m'],
+    usage: '[meta name] [link to meta]',
 
     async execute (message, args) {
         if (!args.length) {
@@ -9,13 +10,24 @@ module.exports = {
             return;
         }
 
+        const metaLink = args.pop();
         const metaName = args.join(' ');
-        const guildManager = message.guild.channels;
 
+        if (!args.length || !metaLink.startsWith("http")) {
+            message.reply("please specify a link.");
+            return;
+        }
+
+        // create category and channels
+        const guildManager = message.guild.channels;
         const category = await guildManager.create(metaName, {type: 'category'});
-        await guildManager.create("🏁" + metaName, {parent: category});
+        const textChannel = await guildManager.create("🏁" + metaName, {parent: category});
         await guildManager.create("🏁" + metaName, {parent: category, type: "voice"});
 
-        await message.delete();
+        // send link and pin
+        const linkMsg = await textChannel.send(`Link: <${metaLink}>`);
+        await linkMsg.pin();
+
+        message.delete();
     }
 };
