+++ commands/Main/puzzle.js
@@ -1,7 +1,8 @@
 module.exports = {
     name: 'puzzle',
     description: `Create a new channel (puzzle) in the current category (meta).`,
-    aliases: ['p'],
+    aliases: ['p', 'problem'],
+    usage: '[puzzle name] [link to puzzle]',
 
     async execute (message, args) {
         if (!args.length) {
@@ -9,11 +10,24 @@ module.exports = {
             return;
         }
 
+        const puzzleLink = args.pop();
         const puzzleName = args.join(' ');
+
+        if (!args.length || !puzzleLink.startsWith("http")) {
+            message.reply("please specify a link.");
+            return;
+        }
+
+        // create channels
         const guildManager = message.guild.channels;
+        const category = message.channel.parent;
+        const textChannel = await guildManager.create("🧩" + puzzleName, {parent: category});
+        await guildManager.create("🧩" + puzzleName, {parent: category, type: "voice"});
 
-        await guildManager.create("🧩" + puzzleName, {parent: message.channel.parent});
+        // send link and pin
+        const linkMsg = await textChannel.send(`Link: <${puzzleLink}>`);
+        await linkMsg.pin();
 
-        await message.delete();
+        message.delete();
     }
 };
